feat: :zap: do first Routes
add MainPage, loader, lazy routes and sharedLayout
