+++ src/modules/Header/Header.jsx
@@ -1,8 +1,8 @@
-import { useState } from "react";
+import { useState } from 'react';
 
-import MenuIcon from "@mui/icons-material/Menu";
+import MenuIcon from '@mui/icons-material/Menu';
 
-import css from "./header.module.css";
+import css from './header.module.css';
 
 const Header = () => {
   const [isShow, setIsShow] = useState(false);
@@ -13,8 +13,8 @@ const Header = () => {
 
   return (
     <header className={css.header}>
-      <div className={css.wrapper + " " + "container"}>
-        <a className={css.logo + " " + "link"} href="#">
+      <div className={css.wrapper + ' ' + 'container'}>
+        <a className={css.logo + ' ' + 'link'} href="#">
           Flower Decor Company
         </a>
         <nav className={css.navigation}>
