+++ src/modules/SharedLayout/SharedLayout.jsx
@@ -1,0 +1,19 @@
+import { Outlet } from "react-router-dom";
+import Header from "../Header";
+import Footer from "../Footer";
+
+// import css from './shared-layout.module.css';
+
+const SharedLayout = () => {
+  return (
+    <div>
+      <Header />
+      <main>
+        <Outlet />
+      </main>
+      <Footer />
+    </div>
+  );
+};
+
+export default SharedLayout;
