--- src/pages/MainPage.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const MainPage = () => {
-  return <div>MainPage</div>;
-};
-
-export default MainPage;
