+++ src/pages/MainPage/main-page.module.css
@@ -1,0 +1,4 @@
+.wrapper {
+  margin: 0;
+  padding: 0;
+}
