+++ src/shared/components/Loader/Loader.jsx
@@ -1,0 +1,22 @@
+import { Rings } from "react-loader-spinner";
+
+import css from "./loader.module.css";
+
+export const Loader = () => {
+  return (
+    <div className={css.loaderWrap}>
+      <Rings
+        height="150"
+        width="150"
+        color="#4fa94d"
+        radius="6"
+        wrapperStyle={{}}
+        wrapperClass=""
+        visible={true}
+        ariaLabel="rings-loading"
+      />
+    </div>
+  );
+};
+
+export default Loader;
