+++ src/shared/components/Loader/index.js
@@ -1,0 +1,1 @@
+export { default } from './Loader';
