+++ src/shared/components/Loader/loader.module.css
@@ -1,0 +1,9 @@
+.loaderWrap {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
